radar-defesa-civil/apps/radar-processor/requirements.txt
wradlib>=2.0.3
arm-pyart>=1.18.0
numpy>=1.26.0
scipy>=1.12.0
xarray>=2024.1.0
h5py>=3.10.0
netCDF4>=1.6.5
redis>=5.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
minio>=7.2.0
Pillow>=10.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
aiofiles>=23.2.0
structlog>=24.1.0
